Fix 'world signup' for logged-in users

+++ app/controllers/registrations_controller.rb
@@ -29,16 +29,6 @@ class RegistrationsController < ApplicationController
   def create
     respond_to do |format|
       format.json do
-        registration_token = self.registration_token or
-          raise "Missing registration token"
-        login_request = LoginRequest
-          .find_by_registration_token!(registration_token)
-        verified_phone_number = login_request.verified_phone_number
-        unless verified_phone_number
-          self.registration_token = nil
-          raise "Phone number not verified"
-        end
-
         user_params = params.expect(user: [
           :name,
           :time_zone,
@@ -52,20 +42,45 @@ class RegistrationsController < ApplicationController
             allow_friend_sharing
           ],
         ])
-        user = User.new(**user_params, phone_number: verified_phone_number)
-        if user.save
-          session.delete(:registration_token)
-          start_new_session_for!(user)
-          render(json: {
-            user: UserSerializer.one(user),
-          })
+        if (user = current_user)
+          if user.update(user_params)
+            render(json: {
+              user: UserSerializer.one(user),
+            })
+          else
+            render(
+              json: {
+                errors: user.form_errors,
+              },
+              status: :unprocessable_content,
+            )
+          end
         else
-          render(
-            json: {
-              errors: user.form_errors,
-            },
-            status: :unprocessable_content,
-          )
+          registration_token = self.registration_token or
+            raise "Missing registration token"
+          login_request = LoginRequest
+            .find_by_registration_token!(registration_token)
+          verified_phone_number = login_request.verified_phone_number
+          unless verified_phone_number
+            self.registration_token = nil
+            raise "Phone number not verified"
+          end
+
+          user = User.new(**user_params, phone_number: verified_phone_number)
+          if user.save
+            session.delete(:registration_token)
+            start_new_session_for!(user)
+            render(json: {
+              user: UserSerializer.one(user),
+            })
+          else
+            render(
+              json: {
+                errors: user.form_errors,
+              },
+              status: :unprocessable_content,
+            )
+          end
         end
       end
     end
